Django 1.9 compatibility changes
